updated webpage: updated access control, updated message sort by date and other forms updated

+++ application/controllers/amessage.php
@@ -4,11 +4,16 @@
  */
 class Amessage extends CI_Controller {
     public function send(){
+    	if(empty($_POST)){
+            redirect();
+    	}
         $this->amessage_model->send();
         redirect(base_url());
     }
 	public function delete(){
+    	if(empty($_POST)){
+            redirect();
+    	}
 		echo $this->amessage_model->delete($this->input->post('amid'));
-
 	}
 }
